fix: Prefab project settings failing to load correctly
Also renamed it to .asset instead of .settings for consistency
