+++ com.unity.netcode.gameobjects/Editor/Configuration/NetcodeForGameObjectsSettings.cs
@@ -2,7 +2,6 @@ using UnityEditor;
 using UnityEngine;
 using UnityEngine.Serialization;
 
-
 namespace Unity.Netcode.Editor.Configuration
 {
     internal class NetcodeForGameObjectsEditorSettings
@@ -38,44 +37,6 @@ namespace Unity.Netcode.Editor.Configuration
         internal static void SetAutoAddNetworkObjectSetting(bool autoAddSetting)
         {
             EditorPrefs.SetBool(AutoAddNetworkObjectIfNoneExists, autoAddSetting);
-        }
-    }
-
-    [FilePath("ProjectSettings/NetcodeForGameObjects.settings", FilePathAttribute.Location.ProjectFolder)]
-    internal class NetcodeForGameObjectsProjectSettings : ScriptableSingleton<NetcodeForGameObjectsProjectSettings>
-    {
-        internal static readonly string DefaultNetworkPrefabsPath = "Assets/DefaultNetworkPrefabs.asset";
-        [SerializeField] public string NetworkPrefabsPath = DefaultNetworkPrefabsPath;
-        public string TempNetworkPrefabsPath;
-
-        private void OnEnable()
-        {
-            if (NetworkPrefabsPath == "")
-            {
-                NetworkPrefabsPath = DefaultNetworkPrefabsPath;
-            }
-            TempNetworkPrefabsPath = NetworkPrefabsPath;
-        }
-
-        [SerializeField]
-        [FormerlySerializedAs("GenerateDefaultNetworkPrefabs")]
-        private byte m_GenerateDefaultNetworkPrefabs;
-
-        public bool GenerateDefaultNetworkPrefabs
-        {
-            get
-            {
-                return m_GenerateDefaultNetworkPrefabs != 0;
-            }
-            set
-            {
-                m_GenerateDefaultNetworkPrefabs = (byte)(value ? 1 : 0);
-            }
-        }
-
-        internal void SaveSettings()
-        {
-            Save(true);
         }
     }
 }
