+++ testproject/ProjectSettings/NetcodeForGameObjects.asset
@@ -9,7 +9,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 2727d53a542a4c1aa312905c3a02d807, type: 3}
   m_Name: 
-  m_EditorClassIdentifier: Unity.Netcode.Editor:Unity.Netcode.Editor.Configuration:NetcodeForGameObjectsProjectSettings
+  m_EditorClassIdentifier: 
+  NetworkPrefabsPath: Assets/DefaultNetworkPrefabs.asset
+  TempNetworkPrefabsPath: Assets/DefaultNetworkPrefabs.asset
   GenerateDefaultNetworkPrefabs: 1
